书签新增修改删除

+++ src/main/java/yj/yajian/photo/controller/FileUploadController.java
@@ -373,7 +373,7 @@ public class FileUploadController {
         return result;
     }
 
-    private void deleteUnuseData(){
+    private void deleteUnuseData() {
         // 遍历 UPLOADED_FOLDER 文件夹下文件，已知文件名作为 dbService map 的 key, dbService 有很多多余数据，如果 key 不在 UPLOADED_FOLDER 里的文件列表中，则删除
         Map<String, Object> all = dbService.getAll();
         // all.forEach((k, v) -> {
@@ -393,7 +393,9 @@ public class FileUploadController {
 
             all.forEach((k, v) -> {
                 if (!existingFiles.contains(k)) {
-                    dbService.remove(k);
+                    if (!k.startsWith("bookmark")) {
+                        dbService.remove(k);
+                    }
                     log.info("RM = {}", k);
                 }
             });
